use only 2 launch, one vanilla, one rviz
